requirements.txt
paho_mqtt==1.6.1
pyserial==3.5
pyttsx3==2.90
requests==2.27.1
